requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
requests==2.31.0
pytest-cov==4.1.0
locust==2.15.1
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.0
responses==0.24.1
pytest-mock==3.12.0
httpx==0.25.1
starlette==0.27.0
